requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
lightgbm
Pillow
